refactor: move profile loading inside profile card
