+++ src/components/Profile/ProfileTippy.tsx
@@ -1,11 +1,9 @@
-import { useQuery } from '@tanstack/react-query';
 import { memo, type PropsWithChildren, useState } from 'react';
 
 import { ProfileCard } from '@/components/Profile/ProfileCard.js';
 import { TippyContext, useTippyContext } from '@/components/TippyContext/index.js';
-import { type SocialSource, Source } from '@/constants/enum.js';
+import { type SocialSource } from '@/constants/enum.js';
 import { Tippy } from '@/esm/Tippy.js';
-import { resolveSocialMediaProvider } from '@/helpers/resolveSocialMediaProvider.js';
 import { useIsMedium } from '@/hooks/useMediaQuery.js';
 import type { Profile } from '@/providers/types/SocialMedia.js';
 
@@ -23,19 +21,8 @@ export const ProfileTippy = memo<ProfileTippyProps>(function ProfileTippy({
     className,
     profile: defaultProfile,
 }) {
-    const [enabled, setEnabled] = useState(false);
     const isMedium = useIsMedium();
-
-    const { data: profile, isLoading } = useQuery({
-        enabled: !!identity && !!source && isMedium && enabled,
-        queryKey: ['profile', source, identity],
-        queryFn: async () => {
-            if (defaultProfile) return defaultProfile;
-            if (!identity || !source) return;
-            const provider = resolveSocialMediaProvider(source);
-            return source === Source.Lens ? provider.getProfileByHandle(identity) : provider.getProfileById(identity);
-        },
-    });
+    const [enabled, setEnabled] = useState(!isMedium);
 
     const insideTippy = useTippyContext();
     if (!isMedium || !children || insideTippy) return children;
@@ -51,12 +38,12 @@ export const ProfileTippy = memo<ProfileTippyProps>(function ProfileTippy({
                 delay={500}
                 arrow={false}
                 trigger="mouseenter"
-                onShow={() => {
+                onTrigger={() => {
                     setEnabled(true);
                 }}
                 hideOnClick
                 interactive
-                content={<ProfileCard profile={profile} loading={isLoading} />}
+                content={enabled ? <ProfileCard source={source} profile={defaultProfile} identity={identity} /> : null}
             >
                 <span className={className}>{children}</span>
             </Tippy>
